backend/app/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
passlib[bcrypt]
pydantic